separate hittable and material
The scene now manages pairs of hittable and material, rather than
requiring hittable instances to track material.

This introduces a new scene::add() method, which adds the hittable and
material to a given scene. Since the scene manages the collection, it's
no longer a requirement that callers create an array of correct size of
hittable pointers.

Finally, destruction is now handled by the scene. Note that since
hittable and material pointers passed into the add() function will be
deleted, neither the original hittable nor the material can be reused
for other object-material pairs -- they must be recreated for a new
pair.